chore: add modular settings files and ADR
