+++ src/game.js
@@ -4,7 +4,7 @@ import { generateRandomCard } from "./helpers/randomCard.js";
 import { DATA_DIR } from "./helpers/constants.js";
 import { shouldReduceMotionSync } from "./helpers/motionUtils.js";
 import { initTooltips } from "./helpers/tooltip.js";
-import { loadSettings } from "./helpers/settingsUtils.js";
+import { loadSettings } from "./helpers/settingsStorage.js";
 import { toggleInspectorPanels } from "./helpers/cardUtils.js";
 import { isEnabled, featureFlagsEmitter } from "./helpers/featureFlags.js";
