+++ tests/helpers/gameRandom.test.js
@@ -28,7 +28,7 @@ describe("game.js", () => {
     const shouldReduceMotionSync = vi.fn(() => true);
     vi.doMock("../../src/helpers/randomCard.js", () => ({ generateRandomCard }));
     vi.doMock("../../src/helpers/motionUtils.js", () => ({ shouldReduceMotionSync }));
-    vi.doMock("../../src/helpers/settingsUtils.js", () => ({
+    vi.doMock("../../src/helpers/settingsStorage.js", () => ({
       loadSettings: vi.fn().mockResolvedValue({ featureFlags: {} })
     }));
