+++ tests/helpers/settingsUtils.test.js
@@ -1,5 +1,6 @@
 import { describe, it, expect, afterEach, vi } from "vitest";
-import { DEFAULT_SETTINGS, resetSettings } from "../../src/helpers/settingsUtils.js";
+import { DEFAULT_SETTINGS } from "../../src/helpers/settingsSchema.js";
+import { resetSettings } from "../../src/helpers/settingsStorage.js";
 
 /**
  * @fileoverview
@@ -38,7 +39,7 @@ describe("settings utils", () => {
    * Should load default settings when localStorage is empty.
    */
   it("loads defaults when storage is empty", async () => {
-    const { loadSettings } = await import("../../src/helpers/settingsUtils.js");
+    const { loadSettings } = await import("../../src/helpers/settingsStorage.js");
     const settings = await loadSettings();
     expect(settings).toEqual(DEFAULT_SETTINGS);
   });
@@ -51,7 +52,7 @@ describe("settings utils", () => {
       "settings",
       JSON.stringify({ featureFlags: { enableTestMode: { enabled: true } } })
     );
-    const { loadSettings } = await import("../../src/helpers/settingsUtils.js");
+    const { loadSettings } = await import("../../src/helpers/settingsStorage.js");
     const settings = await loadSettings();
     expect(settings.featureFlags.randomStatMode).toEqual(
       DEFAULT_SETTINGS.featureFlags.randomStatMode
@@ -64,7 +65,7 @@ describe("settings utils", () => {
    */
   it("saves settings with debounce", async () => {
     vi.useFakeTimers();
-    const { saveSettings } = await import("../../src/helpers/settingsUtils.js");
+    const { saveSettings } = await import("../../src/helpers/settingsStorage.js");
     const data = {
       sound: false,
       motionEffects: true,
@@ -90,7 +91,7 @@ describe("settings utils", () => {
     const original = global.localStorage;
     // Simulate localStorage being undefined
     Object.defineProperty(global, "localStorage", { value: undefined, configurable: true });
-    const { saveSettings, loadSettings } = await import("../../src/helpers/settingsUtils.js");
+    const { saveSettings, loadSettings } = await import("../../src/helpers/settingsStorage.js");
     await expect(saveSettings({ sound: true })).rejects.toThrow();
     await expect(loadSettings()).rejects.toThrow();
     Object.defineProperty(global, "localStorage", { value: original, configurable: true });
@@ -99,7 +100,7 @@ describe("settings utils", () => {
   it("returns defaults if localStorage.getItem returns null", async () => {
     const originalGetItem = Storage.prototype.getItem;
     Storage.prototype.getItem = vi.fn(() => null);
-    const { loadSettings } = await import("../../src/helpers/settingsUtils.js");
+    const { loadSettings } = await import("../../src/helpers/settingsStorage.js");
     const settings = await loadSettings();
     expect(settings).toEqual(DEFAULT_SETTINGS);
     Storage.prototype.getItem = originalGetItem;
@@ -110,7 +111,7 @@ describe("settings utils", () => {
    */
   it("updates a single setting and persists", async () => {
     vi.useFakeTimers();
-    const { updateSetting, loadSettings } = await import("../../src/helpers/settingsUtils.js");
+    const { updateSetting, loadSettings } = await import("../../src/helpers/settingsStorage.js");
     const promise = updateSetting("sound", false);
     await vi.runAllTimersAsync();
     await promise;
@@ -123,7 +124,7 @@ describe("settings utils", () => {
    */
   it("recovers from invalid stored JSON", async () => {
     localStorage.setItem("settings", "{bad json}");
-    const { loadSettings } = await import("../../src/helpers/settingsUtils.js");
+    const { loadSettings } = await import("../../src/helpers/settingsStorage.js");
     const warnSpy = vi.spyOn(console, "warn").mockImplementation(() => {});
     const settings = await loadSettings();
     expect(settings).toEqual(DEFAULT_SETTINGS);
@@ -138,7 +139,7 @@ describe("settings utils", () => {
     Storage.prototype.setItem = vi.fn(() => {
       throw new Error("fail");
     });
-    const { saveSettings } = await import("../../src/helpers/settingsUtils.js");
+    const { saveSettings } = await import("../../src/helpers/settingsStorage.js");
     await expect(
       saveSettings({
         sound: true,
@@ -164,7 +165,7 @@ describe("settings utils", () => {
    */
   it("rejects when updating unknown key", async () => {
     vi.useFakeTimers();
-    const { updateSetting, loadSettings } = await import("../../src/helpers/settingsUtils.js");
+    const { updateSetting, loadSettings } = await import("../../src/helpers/settingsStorage.js");
     await expect(updateSetting("nonexistentKey", "value")).rejects.toThrow(
       "Schema validation failed"
     );
@@ -179,7 +180,7 @@ describe("settings utils", () => {
    */
   it("debounces multiple saveSettings calls", async () => {
     vi.useFakeTimers();
-    const { saveSettings } = await import("../../src/helpers/settingsUtils.js");
+    const { saveSettings } = await import("../../src/helpers/settingsStorage.js");
     const data1 = {
       sound: true,
       motionEffects: true,
@@ -218,7 +219,7 @@ describe("settings utils", () => {
 
   it("resets settings to defaults", async () => {
     localStorage.setItem("settings", JSON.stringify({ sound: true }));
-    const { resetSettings } = await import("../../src/helpers/settingsUtils.js");
+    const { resetSettings } = await import("../../src/helpers/settingsStorage.js");
     const result = resetSettings();
     expect(result).toEqual(DEFAULT_SETTINGS);
     expect(JSON.parse(localStorage.getItem("settings"))).toEqual(DEFAULT_SETTINGS);
@@ -226,7 +227,8 @@ describe("settings utils", () => {
 
   it("provides synchronous access via getSetting", async () => {
     localStorage.setItem("settings", JSON.stringify({ sound: false }));
-    const { loadSettings, getSetting } = await import("../../src/helpers/settingsUtils.js");
+    const { loadSettings } = await import("../../src/helpers/settingsStorage.js");
+    const { getSetting } = await import("../../src/helpers/settingsCache.js");
     await loadSettings();
     expect(getSetting("sound")).toBe(false);
     localStorage.setItem("settings", JSON.stringify({ sound: true }));
@@ -238,7 +240,8 @@ describe("settings utils", () => {
       "settings",
       JSON.stringify({ featureFlags: { enableTestMode: { enabled: true } } })
     );
-    const { loadSettings, getFeatureFlag } = await import("../../src/helpers/settingsUtils.js");
+    const { loadSettings } = await import("../../src/helpers/settingsStorage.js");
+    const { getFeatureFlag } = await import("../../src/helpers/settingsCache.js");
     await loadSettings();
     expect(getFeatureFlag("enableTestMode")).toBe(true);
     expect(getFeatureFlag("nonexistent")).toBe(false);
